requirements.txt
argon2-cffi==19.1.0
Babel==2.6.0
backports.ssl-match-hostname==3.7.0.1
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cloudinary==1.20.0
construct==2.5.3
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.2
django-appconf==1.0.3
django-bootstrap3==11.0.0
django-braces==1.13.0
django-ckeditor==5.7.0
django-compressor==2.3
django-environ==0.4.5
django-extra-views==0.11.0
django-filter==2.1.0
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-mptt==0.10.0
django-pagedown==1.0.6
django-phonenumber-field==2.0.1
django-summernote==0.8.11.6
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
factory-boy==2.11.1
Faker==1.0.2
future==0.17.1
idna==2.8
iso8601==0.1.12
jsonpickle==1.2
Markdown==3.1.1
markdown2==2.3.7
mock==2.0.0
pbr==5.1.3
pefile==2019.4.18
phonenumbers==8.10.10
Pillow==5.4.1
purl==1.5
pycountry==18.12.8
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.7.5
pytz==2018.9
rcssmin==1.0.6
requests==2.21.0
rjsmin==1.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
Unidecode==1.0.23
urllib3==1.24.1
validators==0.13.0
wincertstore==0.2
